Refactor database setup into its own project; remove clunky port checking mechanism
