+++ CSETWebApi/CSETWeb_Api/CSETWebCore.DatabaseManager/CSETWebCore.DatabaseManager.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.1" />
+  </ItemGroup>
+
 </Project>
