--- CSETWebApi/CSETWeb_Api/CSETWebCore.DatabaseManager/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace CSETWebCore.DatabaseManager
-{
-    public class Class1
-    {
-    }
-}
