+++ CSETWebApi/CSETWeb_Api/CSETWeb_ApiCore/CSETWebCore.Api.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
+    <AssemblyVersion>11.0.0.0</AssemblyVersion>
   </PropertyGroup>
 
   <ItemGroup>
@@ -21,6 +22,7 @@
   <ItemGroup>
     <ProjectReference Include="..\CSETWebCore.Business\CSETWebCore.Business.csproj" />
     <ProjectReference Include="..\CSETWebCore.CryptoBuffer\CSETWebCore.CryptoBuffer.csproj" />
+    <ProjectReference Include="..\CSETWebCore.DatabaseManager\CSETWebCore.DatabaseManager.csproj" />
     <ProjectReference Include="..\CSETWebCore.DatabaseUpgrade\CSETWebCore.DatabaseManager.csproj" />
     <ProjectReference Include="..\CSETWebCore.DataLayer\CSETWebCore.DataLayer.csproj" />
     <ProjectReference Include="..\CSETWebCore.ExportCSV\CSETWebCore.ExportCSV.csproj" />
